JHA: stepper refactor (WIP).
