+++ jha/src/App.vue
@@ -4,13 +4,6 @@
       <Header :title='pageTitle'
               imagePath='/jha/images/' />
       <main>
-        <div class="container stepper">
-          <PageStepper :currentPath='$router.currentRoute.path'
-                      :routes='stepRoutes'
-                      @toggleShowMobileDetails='handleToggleShowMobileStepperDetails($event)'
-                      :isMobileStepperOpen='isMobileStepperOpen'
-                      @onClickLink='handleClickStepperLink($event)'/>
-        </div>
         <router-view/>
       </main>
       <Footer :version='version' />
@@ -28,22 +21,13 @@ import project from '../package.json';
 import {
   Header,
   Footer,
-  PageStepper,
 } from 'common-lib-vue';
-import {
-  MODULE_NAME as appModule,
-  SET_SHOW_MOBILE_STEPPER_DETAILS,
-} from '@/store/modules/app-module';
 import {
   MODULE_NAME as enrolmentModule,
   SET_APPLICATION_UUID,
 } from '@/store/modules/enrolment-module';
 import pageStateService from '@/services/page-state-service';
-import { scrollTo } from '@/helpers/scroll';
-import {
-  commonRoutes,
-  enrolmentRoutes,
-} from './router/routes';
+import { commonRoutes } from '@/router/routes';
 import { Wormhole } from 'portal-vue';
 import { v4 as uuidv4 } from 'uuid';
 import spaEnvService from '@/services/spa-env-service';
@@ -54,7 +38,6 @@ export default {
   components: {
     Header,
     Footer,
-    PageStepper,
   },
   data: () => {
     return {
@@ -86,28 +69,6 @@ export default {
       });
   },
   computed: {
-    stepRoutes() {
-      const routes = [
-        {...enrolmentRoutes.FORM_SELECTION_PAGE},
-        {...enrolmentRoutes.PERSONAL_INFO_PAGE},
-        {...enrolmentRoutes.SPOUSE_INFO_PAGE},
-        {...enrolmentRoutes.CHILD_INFO_PAGE},
-      ];
-      if (this.$store.state.enrolmentModule.isApplyingForFPCare) {
-        routes.push({...enrolmentRoutes.FPCARE_INFO_PAGE});
-      }
-      if (this.$store.state.enrolmentModule.isApplyingForSuppBen) {
-        routes.push({...enrolmentRoutes.SUPP_BEN_INFO_PAGE});
-        routes.push({...enrolmentRoutes.DOCUMENTS_PAGE});
-      }
-      routes.push({...enrolmentRoutes.CONTACT_INFO_PAGE});
-      routes.push({...enrolmentRoutes.REVIEW_PAGE});
-      routes.push({...enrolmentRoutes.SUBMISSION_PAGE});
-      return routes;
-    },
-    isMobileStepperOpen() {
-      return this.$store.state.appModule.showMobileStepperDetails;
-    },
     isModalOpen() {
       const modalTargetEl = document.body.querySelector('#modal-target');
       const modalTargetHasChildren = modalTargetEl && modalTargetEl.children.length > 0;
@@ -115,27 +76,11 @@ export default {
           || modalTargetHasChildren;
     }
   },
-  methods: {
-    handleToggleShowMobileStepperDetails(isDetailsShown) {
-      this.$store.dispatch(appModule + '/' + SET_SHOW_MOBILE_STEPPER_DETAILS, isDetailsShown);
-    },
-    handleClickStepperLink(path) {
-      pageStateService.setPageIncomplete(this.$router.currentRoute.path);
-      pageStateService.setPageComplete(path);
-      this.$router.push(path);
-      scrollTo(0);
-    }
-  }
 }
 </script>
 
 <style scoped>
 main {
   padding: 0;
-}
-@media only screen and (max-width: 575px) {
-  .container.stepper {
-    padding: 0;
-  }
 }
 </style>
