+++ jha/src/styles/styles.css
@@ -36,3 +36,9 @@ body.no-scroll {
 .field-description {
   font-size: 13.33px;
 }
+
+@media only screen and (max-width: 575px) {
+  .container.stepper {
+    padding: 0;
+  }
+}
